Add starter-for-ten process for managing paas org
We've had some issues recently with our shared PaaS org around user
management and memory quotas.  This adds some documentation to
describe some processes in these areas.

I'm not sure that this is the best place for the docs to live, but we
don't currently have an obvious home for RE-wide internal
documentation.