Update grafana content to reflect current practices
- It is rare that teams are using the GDS default dashboard because
they are now creating their own dashboards and using those
- We also have other sources of metrics meaning teams may also
use the container metrics dashboard or metrics for tools like
elasticsearch.

Based on this I have:

- Added a short explanation for how team folders are organised
- Added definitions of the purpose of the dashboards in the
General folder
- Moved the detailed explanation of the default application dashboard
into it's own page (but have made no content changes)
- Removed some of the simpler content as I don't think we have seen
evidence of users struggling to use Grafana so far

Note: I have kept the section on team dashboard permissions brief
and vague - essentially contact us for help - as I think we need
to do more on team permissions. At the moment, teams will not
be able to manage the permissions for users of their team without
having admin privileges of the whole system as per
https://github.com/grafana/grafana/issues/11584. This may be something
we want to investigate further to avoid us needing to manage all
user permissions in terms of which teams they can access and instead
being able to put this burden on the teams themselves.