Make distinction between application and container level metrics
We've seen in user research users confused about the difference
between our exporters. This is a small tweak to hopefully make that
less confusing. We could likely make more dramatic changes to the
index page but I am going for speed and least controversy with
this PR in the hopes that it makes small immediate benefit.