Add further annonations to our example alert
The aim is by adding in these annonations we will encourage
users to do similar when writing their own alerts and therefore
provide sufficient context for anyone needing to investigate an
alert.

We are looking to test this in user research so this is just a
first pass that hasn't had too much thought.