Update AWS access guidance to recommend aws-vault
The current guidance is problematic for a number of reasons because it
requires storing AWS keys unencrypted in a file in your home directory.
It's also fiddly to use it with tools other than the AWS CLI.

I've evaluated a few approaches to solving these issues, and determined
that aws-vault is the best option for our needs. This therefore updates
the docs to recommend its usage.

Other avenues explored:

As part of this assessment I looked at several approaches for handling
the keys with the AWS CLI that didn't require storing them in the
credentials file:

* Using `credential_source = Environment` in the profile config. I
  couldn't get this to work - the AWS CLI seems to ignore creds in the
  environment when the `--profile` option is given, adding this setting to
  the config doesn't help.

* Using the `credential_process = ...` option[1]. This allows shelling
  out to a process that returns some JSON, which would allow sourcing the
  credentials from something like `pass`. This works for direct access to
  an account, but when used in the source account for an assume-role
  profile, the CLI returned the following error:

> The source_profile "gds-users" must specify either static credentials
> or an assume role configuration

Even if these options did work it would still be necessary to add some
scripting to allow an assume-role profile to be easily used with tools
like terraform. It might also be possible to work around some of the
above problems with some scripting, but at this point we start to
reimplement aws-vault... in bash... badly...

Given aws-vault has already solved these issues, and has a smooth
workflow, it makes sense to recommend it.

[1]https://docs.aws.amazon.com/cli/latest/topic/config-vars.html#sourcing-credentials-from-external-processes