Remove unneeded files
The deploy script seemed to be specific to the GOV.UK service
manual and we are just using Travis to deploy now.

The nginx file does not appear to be used and I can't see it
included in other repo's that use the tech docs template (for
example the GDS Way) so believe we do not need it either.