remove logit mutual auth section
We don't think many people use this.  Logit don't have a particularly
consistent process here in our experience.  Remove this to keep the
manual tidier and easier to understand for the common use cases.

As a result, I've simplified the Filebeat section to remove the
references to mutual auth.  I also removed the puppet configuration to
keep things simple.