Proof of concept for reordered content
Our instructions for setting up the client libraries and
paas-prometheus-exporter, protecting it and binding it to
Prometheus had slowly iterated into what I thought was a slightly
confusing order. Content was duplicated and instructions in
some sections had content that was to be done by both client
libraries and the paas-prometheus exporter whereas other
sections were only for one type but this was not made
explicit. In particular the IP whitelisting was in both the
paas-prometheus-exporter set up section but also the bind your
exporter to Prometheus section (but only for client libraries).

This is a proof of concept of what it could look like if we took
a more linear approach to our docs.

This commit at the moment isn't perfect, some of the wording
could be more precise and some things could be made into links.
May also be beneficial to get the eyes of a tech writer on this.