Cross-reference re-team-manual on the reliability-engineering home page
We have a number of different places for documentation that are unlikely to be
consolidated. In light of this it makes sense to cross-reference where possible
so that documentation need not be repeated again and again and it's possible to
find -all- most of the resources if you have found one.

The re-team-manual has an existing reference for the reliability-engineering
docs so this commit balances that out and completes the cross-reference between
those two sites.