Remove redundant pages and mark as deprecated
I've removed some of the quick wins such as talking about RE strategy.

I've kept most of the technical documentation as this should be
reviewed with a finer eye as some of it may still be useful. The
exception to this is the monitoring and alerting section as I think
this has little use by GDS generally.